+++ tests/SmscRuChannelTest.php
@@ -3,11 +3,11 @@
 namespace NotificationChannel\SmscRu\Tests;
 
 use Mockery as M;
+use Illuminate\Notifications\Notifiable;
 use Illuminate\Notifications\Notification;
 use NotificationChannels\SmscRu\SmscRuApi;
 use NotificationChannels\SmscRu\SmscRuChannel;
 use NotificationChannels\SmscRu\SmscRuMessage;
-use NotificationChannels\SmscRu\Exceptions\CouldNotSendNotification;
 
 class SmscRuChannelTest extends \PHPUnit_Framework_TestCase
 {
@@ -87,17 +87,36 @@ class SmscRuChannelTest extends \PHPUnit_Framework_TestCase
     /** @test */
     public function it_does_not_send_a_message_when_to_missed()
     {
-        $this->expectException(CouldNotSendNotification::class);
+        $this->smsc->shouldNotReceive('send');
 
         $this->channel->send(
             new TestNotifiableWithoutRouteNotificationForSmscru(), new TestNotification()
         );
     }
+
+    /** @test */
+    public function it_can_send_a_notification_to_multiple_phones()
+    {
+        $this->smsc->shouldReceive('send')->once()
+            ->with(
+                [
+                    'phones'  => '+1234567890,+0987654321,+1234554321',
+                    'mes'     => 'hello',
+                    'sender'  => 'John_Doe',
+                ]
+            );
+
+        $this->channel->send(new TestNotifiableWithManyPhones(), new TestNotification());
+    }
 }
 
 class TestNotifiable
 {
-    public function routeNotificationFor()
+    use Notifiable;
+
+    // Laravel v5.6+ passes the notification instance here
+    // So we need to add `Notification $notification` argument to check it when this project stops supporting < 5.6
+    public function routeNotificationForSmscru()
     {
         return '+1234567890';
     }
@@ -105,9 +124,17 @@ class TestNotifiable
 
 class TestNotifiableWithoutRouteNotificationForSmscru extends TestNotifiable
 {
-    public function routeNotificationFor()
+    public function routeNotificationForSmscru()
     {
         return false;
+    }
+}
+
+class TestNotifiableWithManyPhones extends TestNotifiable
+{
+    public function routeNotificationForSmscru()
+    {
+        return ['+1234567890', '+0987654321', '+1234554321'];
     }
 }